Inherit common channel methods
